Added unmount check - to avoid update on umounted component
We need to make sure that the component was not unmounted before loadScript completes loading remote script and triggering its callback.

**Note**

I had some issue with the **lint**. When I run `yarn lint:fix in my local machine, it has replaced (replaced . with <=, i.e replaced space with enter for below lines of codes.
```
        auth2.grantOfflineAccess(options).then(res => onSuccess(res), err => onFailure(err))

```
with
```
	auth2.grantOfflineAccess(options).then(
          res => onSuccess(res),
          err => onFailure(err)
        )
```

But once I submitted a PR, it is throwing an lint check error, mentioning that need to replace <= with . (replace enter with spaces)

What could be causing this problem ?